third_party/WebKit/LayoutTests/fast/encoding/api/sharedarraybuffer.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>

if (window.SharedArrayBuffer) {
  test(() => {
    const decoder = new TextDecoder('utf-8');
    assert_throws(null, () => {
      decoder.decode(new Uint8Array(new SharedArrayBuffer(4)));
    }, 'constructing TextDecoder with SharedArrayBuffer view should throw');
  }, 'decoding SharedArrayBuffer');
}
done();

</script>
